requirements.txt
pdfplumber>=0.9.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
transformers>=4.35.0
accelerate>=0.24.0
torch>=2.0.0
streamlit>=1.28.0
requests>=2.31.0
django>=5.0.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
tqdm>=4.66.0
